requirements.txt
pytesseract
pyautogui
Pillow
mss